[ARM] msm: trout/sapphire: move *_CPLD_BASE outside the vmalloc range
SAPPHIRE_CPLD_BASE and TROUT_CPLD_BASE were inside the range of memory
allocated by the kernel for virtual mappings, causing the IO mapping
to be overwritten if the kernel allocated the space.

To reproduce, allocate then deallocate all virtual kernel memory.
The mapping will be removed and the kernel will crash when any of
the IO memory is accessed (e.g. for changing phone brightness).

Change-Id: I74d0afa61535ae1ce27c7d1332ff0155af860799
Signed-off-by: Christopher Lais <chris+android@zenthought.org>
Signed-off-by: Larry Bassel <lbassel@codeaurora.org>